+++ src/main/java/org/hv/biscuits/core/views/PermissionView.java
@@ -1,13 +1,19 @@
 package org.hv.biscuits.core.views;
 
+import java.io.Serializable;
+
 /**
  * @author leyan95
  */
-public class PermissionView {
-    private final String id;
+public class PermissionView implements Serializable {
+    private static final long serialVersionUID = -3750247801987645175L;
+    private String id;
     private String bundleId;
-    private final String name;
-    private final String comment;
+    private String name;
+    private String comment;
+
+    public PermissionView() {
+    }
 
     private PermissionView(String id, String name, String comment) {
         this.id = id;
@@ -19,6 +25,14 @@ public class PermissionView {
         return new PermissionView(id, name, comment);
     }
 
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
     public String getBundleId() {
         return bundleId;
     }
@@ -27,15 +41,19 @@ public class PermissionView {
         this.bundleId = bundleId;
     }
 
-    public String getId() {
-        return id;
-    }
-
     public String getName() {
         return name;
     }
 
+    public void setName(String name) {
+        this.name = name;
+    }
+
     public String getComment() {
         return comment;
+    }
+
+    public void setComment(String comment) {
+        this.comment = comment;
     }
 }