+++ src/main/java/org/hv/biscuits/spine/identify/ServerTableStationDateGenerator.java
@@ -1,12 +1,11 @@
 package org.hv.biscuits.spine.identify;
 
-import org.hv.biscuits.spine.model.TableIdView;
+import org.hv.biscuits.spine.model.TableNameId;
 import org.hv.pocket.criteria.Criteria;
 import org.hv.pocket.criteria.Restrictions;
 import org.hv.pocket.identify.AbstractIdentifyGenerator;
 import org.hv.pocket.model.AbstractEntity;
 import org.hv.pocket.model.MapperFactory;
-import org.hv.pocket.query.SQLQuery;
 import org.hv.pocket.session.Session;
 import org.hv.pocket.session.SessionFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -75,13 +74,13 @@ public class ServerTableStationDateGenerator extends AbstractIdentifyGenerator {
         Session session = SessionFactory.getSession("biscuits");
         session.open();
         String tableName = MapperFactory.getTableName(clazz.getName());
-        SQLQuery query = session.createSQLQuery("SELECT T.TABLE_NAME AS tableName, T.TABLE_ID AS tableId FROM T_TABLE_NAME_ID T WHERE T.TABLE_NAME = :TABLE_NAME", TableIdView.class)
-                .setParameter("TABLE_NAME", tableName);
-        TableIdView tableIdView = (TableIdView) query.unique();
+        Criteria criteria = session.createCriteria(TableNameId.class);
+        criteria.add(Restrictions.equ("tableName", tableName));
+        TableNameId tableNameId = criteria.unique();
         session.close();
-        if (tableIdView == null) {
+        if (tableNameId == null) {
             throw new NullPointerException(String.format("数据库表T_TABLE_NAME_ID中未找到TABLE_NAME为%s的数据", tableName));
         }
-        return tableIdView.getTableId();
+        return tableNameId.getTableId();
     }
 }